refactor gemm core
