+++ autotune/tune/utils.py
@@ -12,7 +12,7 @@ from neuronxcc.nki.compile import GenericKernel
 
 from autotune.cache.directories import split_file_info
 from autotune.core.lhs_rhs import gemm_main
-from autotune.core.matmul import matmul_main
+from autotune.core.lhsT_rhs import matmul_main
 from autotune.golden.gemm import matmul_op, matmul_xt_op
 from autotune.golden.rmsnorm_linear import rmsnorm_gemm_golden
 from autotune.tune.metrics import extract_metrics
